pyproject.toml
[project]
name = "rss"
version = "0.1.0"
description = "MCP Server to fetch RSS feeds and return them as markdown"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "fastmcp>=2.11.3",
    "feedparser>=6.0.11",
    "pyyaml>=6.0.2",
]
